# docs/swagger/generated.gen.swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: hubme api
servers:
  - url: http://localhost:8080
paths:
...
  /community/{uuid}:
    get:
      summary: コミュニティ情報取得
      description: コミュニティ情報をuuidで取得
      operationId: CommunityFindByID
      parameters:
        - $ref: '#/components/parameters/CommunityUUIDRequest'
      responses:
        '200':
          description: Successfully get community data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: successful get community by id
                  user:
                    $ref: '#/components/schemas/Community'
        '400':
          description: Failed to find community data by uuid
    put:
      summary: コミュニティ情報を更新
      description: コミュニティ情報をuuidで更新
      operationId: CommunityFindByID
      parameters:
        - $ref: '#/components/parameters/CommunityUUIDRequest'
      responses:
        '200':
          description: Successfully get community data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: successful get community by id
                  user:
                    $ref: '#/components/schemas/Community'
        '400':
          description: Failed to find community data by uuid
...
components:
  parameters:
...
    CommunityUUIDRequest:
      name: uuid
      in: path
      required: true
      description: コミュニティのID
      schema:
        type: string
  schemas:
...
    Community:
      type: object
      properties:
        uuid:
          type: string
          format: string
          description: コミュニティのuuid
          example: ffafafafffafaf
        name:
          type: string
          format: string
          description: ”コミュニティ名"
          example: RCC
        email:
          type: string
          format: string
          description: メールアドレス
          example: email
        password:
          type: string
          format: string
          description: パスワード
          example: hogehoge
        img:
          type: string
          format: string
          description: イメージのs3パス
          example: hogehoge.jpeg
        self:
          type: string
          format: string
          description: 自己紹介
          example: hogehoge
        mem1:
          type: string
          format: string
          description: 所属1
          example: A univ.
        mem2:
          type: string
          format: string
          description: 所属２
          example: B Club
        mem3:
          type: string
          format: string
          description: 所属3
          example: ''
      required:
        - uuid
        - name
        - email
        - password
...